Safari-609.1.11.1/LayoutTests/crypto/subtle/hmac-import-raw-key-export-raw-key-expected.txt

Test importing a raw HMAC key with SHA-1 and then export it in raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS exportedRawKey is rawKey
PASS successfullyParsed is true

TEST COMPLETE
